chore: move http related constants into api/utils/httpConsts.ts
