+++ app/api/errors/index.ts
@@ -1,21 +1,4 @@
-/**
- * Whenever someone needs to reference new status code,
- * please add here
- * https://developer.mozilla.org/en-US/docs/Web/HTTP/Status
- */
-export enum HttpStatusCode {
-  OK = 200,
-  Created = 201,
-  NoContent = 204,
-  BadRequest = 400,
-  UnAuthorized = 401,
-  Forbidden = 403,
-  NotFound = 404,
-  MethodNotAllowed = 405,
-  IamAteapot = 418,
-  InternalServerError = 500,
-  NotImplemented = 501
-}
+import { HttpStatusCode } from '../utils/httpConsts'
 
 /**
  * This class standardizes the error object returned from server
